backend/requirements.txt
fastapi==0.110.0
uvicorn==0.25.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.1
cryptography==42.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.7
email-validator==2.1.0
alembic==1.13.1